+++ db/migrate/20190221202712_add_primary_to_projects.rb
@@ -1,0 +1,6 @@
+class AddPrimaryToProjects < ActiveRecord::Migration[6.0]
+  def change
+    add_column :projects, :primary, :boolean, null: false, default: false
+    add_index :projects, :primary
+  end
+end