+++ CHANGELOG.md
@@ -5,6 +5,9 @@
   - Landing page survey count now includes newly completed surveys from Slice
     along with original survey count
   - Tabs now display better on smaller screens
+- **Project Changes**
+  - A primary project can now be specified that determines the consent displayed
+    on the root consent URL
 - **Survey Changes**
   - Dashboard now displays next survey question
 - **Gem Changes**